predictive-analytics-service/requirements.txt
paho-mqtt
statsmodels